% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{create_server}
\alias{create_server}
\title{Create server for the Rapids Shiny App}
\usage{
create_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
None. Sets up Shiny outputs and observers.
}
\description{
This function sets up the server logic for the RapidsShiny app.
}
