% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{create_ui}
\alias{create_ui}
\title{Create UI for the Rapids Shiny App}
\usage{
create_ui()
}
\value{
A shiny dashboardPage object
}
\description{
This function generates the dashboardPage UI for the RapidsShiny app.
}
